% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharperatio.R
\name{data_import}
\alias{data_import}
\title{Price data of the stock}
\usage{
data_import(x)
}
\arguments{
\item{x}{Symbol of stock}
}
\value{
price data of the stock
}
\description{
Get data from the package
}
