% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dseprojectS.R
\name{stock_plot_return}
\alias{stock_plot_return}
\title{Plot and Simpe return}
\usage{
stock_plot_return(symbol)
}
\arguments{
\item{symbol}{Symbol of stock}
}
\description{
Plot the open, the close and the adjusted price of the stock and give a simple stock return.
}
