% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharperatio.R
\name{stock_simple_return}
\alias{stock_simple_return}
\title{Simple return of the stock}
\usage{
stock_simple_return(x)
}
\arguments{
\item{x}{Symbol of stock}
}
\value{
stock return rate of the stock
}
\description{
Calculate the simple return of the stock
}
