% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharperatio.R
\name{sharpe_ratio}
\alias{sharpe_ratio}
\title{Sharpe Ratio of selected stock}
\usage{
sharpe_ratio(x)
}
\arguments{
\item{x}{Symbol of stock}
}
\value{
stock return rate of the latest date
}
\description{
Calculate the Sharpe Ratio of selected stock
}
