% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharperatio.R
\name{stock_plot_return}
\alias{stock_plot_return}
\title{Plot and Simpe return}
\usage{
stock_plot_return(x)
}
\arguments{
\item{x}{Symbol of stock}
}
\value{
plot of the chosen stock
}
\description{
Plot the open, the close and the adjusted price of the stock and give a simple stock return.
}
