% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dseprojectS.R
\name{stock_simple_return}
\alias{stock_simple_return}
\title{Plot and Simpe return}
\usage{
stock_simple_return(x)
}
\arguments{
\item{x}{Symbol of stock}
}
\value{
stock return rate of the latest date
}
\description{
Plot the open, the close and the adjusted price of the stock and give a simple stock return.
}
